Refactored of code in all activities

+++ app/src/main/java/com/applocum/connecttomyhealth/ui/medicalhistory/models/Medical.kt
@@ -5,7 +5,7 @@ import java.io.Serializable
 
 data class Medical(
     @SerializedName("id") val id: Int,
-    @SerializedName("code") val code: Int,
+    @SerializedName("code") val code: Long,
     @SerializedName("description") val description: String,
     @SerializedName("full_code") val full_code: String
 ) : Serializable
